Refactor: Separate profile creation component
Separate the manual profile entry form into a dedicated component. This will allow for better organization and potential future enhancements, such as a separate upload feature.
